Favors a Rails Concern over adding logic to the policy
This way, we can target the impacted controllers, specifically.
