+++ app/controllers/v0/profile/alternate_phones_controller.rb
@@ -4,8 +4,9 @@ module V0
   module Profile
     class AlternatePhonesController < ApplicationController
       include Vet360::Writeable
+      include Authorizeable
 
-      before_action { authorize :evss, :access? }
+      before_action :authorize_evss!
 
       def show
         response = service.get_alternate_phone
