+++ app/controllers/v0/profile/primary_phones_controller.rb
@@ -4,8 +4,9 @@ module V0
   module Profile
     class PrimaryPhonesController < ApplicationController
       include Vet360::Writeable
+      include Authorizeable
 
-      before_action { authorize :evss, :access? }
+      before_action :authorize_evss!
 
       def show
         response = service.get_primary_phone
