+++ app/policies/evss_policy.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-# rubocop:disable Metrics/BlockLength
 EVSSPolicy = Struct.new(:user, :evss) do
   include SentryLogging
 
@@ -20,26 +19,7 @@ EVSSPolicy = Struct.new(:user, :evss) do
         profile: 'pciu_profile'
       )
 
-      raise Common::Exceptions::Forbidden.new(detail: error_detail, source: 'EVSS')
+      false
     end
   end
-
-  def error_detail
-    "User does not have access to the requested resource due to missing values: #{missing_values}"
-  end
-
-  # Returns a comma-separated string of the user's blank attributes. `participant_id` is AKA `corp_id`.
-  #
-  # @return [String] Comma-separated string of the attribute names
-  #
-  def missing_values
-    missing = []
-
-    missing << 'corp_id' if user.participant_id.blank?
-    missing << 'edipi' if user.edipi.blank?
-    missing << 'ssn' if user.ssn.blank?
-
-    missing.join(', ')
-  end
 end
-# rubocop:enable Metrics/BlockLength
